requirements.txt
cffi==1.2.1
characteristic==14.3.0
cryptography==1.0.2
cssselect==0.9.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-daterange-filter==1.2.0
django-enumfield==1.2.1
django-enumfields==0.7.4
django-toolbelt==0.0.1
enum34==1.0.4
gunicorn==19.3.0
idna==2.0
ipaddress==1.0.14
lxml==3.4.4
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
pytz==2015.6
queuelib==1.4.2
Scrapy==1.0.3
service-identity==14.0.0
six==1.9.0
static3==0.6.1
Twisted==15.4.0
w3lib==1.12.0
wheel==0.24.0
zope.interface==4.1.2
